+++ .eslintrc.json
@@ -37,6 +37,7 @@
     "import/no-unresolved": "off",
     "import/extensions": "off",
     "import/prefer-default-export": "off",
+    "react/jsx-props-no-spreading": "off",
     "no-param-reassign": [
       "error",
       {
