+++ src/pages/home-page/ui/HomePage.module.scss
@@ -1,22 +1,19 @@
+.layout {
+  display: flex;
+  width: 100%;
+}
+
 .content {
   display: grid;
   grid-template-rows: auto 1fr auto;
-
   width: 50%;
   height: 100dvh;
-  padding: 32px 64px;
 
   @media (max-width: 1439px) {
     width: 57%;
-    padding: 32px 36px 32px 32px;
   }
 
   @media (max-width: 1023px) {
     width: 100%;
-  }
-
-  @media (max-width: 767px) {
-    width: 100%;
-    padding: 0;
   }
 }
