--- src/pages/home-page/ui/HomePage/HomePage.jsx
@@ -1,13 +1,0 @@
-import { HomePageSlider } from '@widgets/homepage-slider'
-
-import HomePageContent from '../HomePageContent/HomePageContent'
-import { layout } from './HomePage.module.scss'
-
-export function HomePage() {
-  return (
-    <div className={layout}>
-      <HomePageContent />
-      <HomePageSlider />
-    </div>
-  )
-}
