--- src/pages/home-page/ui/HomePage/HomePage.module.scss
@@ -1,4 +1,0 @@
-.layout {
-  display: flex;
-  width: 100%;
-}
