+++ src/shared/ui/icon/index.js
@@ -1,1 +1,1 @@
-export { Icon } from './ui/Icon'
+export { Icon } from './ui/Icon.jsx'
