+++ src/widgets/footer/ui/Footer.module.scss
@@ -1,6 +1,7 @@
 .footer {
   display: flex;
   justify-content: space-between;
+  padding: 32px 64px;
 
   span {
     font-size: 13px;
@@ -24,6 +25,10 @@
     @media (max-width: 767px) {
       color: var(--c-green);
     }
+  }
+
+  @media (max-width: 1439px) {
+    padding: 32px 36px 32px 32px;
   }
 
   @media (max-width: 767px) {
