+++ src/widgets/header/ui/Header.module.scss
@@ -2,11 +2,16 @@
   display: flex;
   align-items: center;
   justify-content: space-between;
+  padding: 32px 64px;
+
+  @media (max-width: 1439px) {
+    padding: 32px 36px 32px 32px;
+  }
 
   @media (max-width: 767px) {
     flex-direction: column;
     gap: 8px;
     align-items: flex-end;
-    padding: 16px 16px 0;
+    padding: 16px;
   }
 }
