+++ src/widgets/header/ui/Logo.jsx
