+++ src/widgets/header/ui/Logo.module.scss
