+++ src/widgets/homepage-main/index.js
@@ -1,0 +1,1 @@
+export { HomePageMain } from './ui/HomePageMain.jsx'
