+++ src/widgets/modal/ui/Modal.module.scss
@@ -22,6 +22,16 @@
 }
 
 .modal {
+  button {
+    padding: 0;
+
+    font: inherit;
+    color: inherit;
+
+    background-color: transparent;
+    border: none;
+  }
+
   @media (max-width: 767px) {
     position: absolute;
     top: 16px;
@@ -30,19 +40,23 @@
 }
 
 .icon-open {
-  @extend %icon;
+  svg {
+    @extend %icon;
 
-  @media (max-width: 767px) {
-    color: var(--c-black);
-
-    &:active {
+    @media (max-width: 767px) {
       color: var(--c-black);
+
+      &:active {
+        color: var(--c-black);
+      }
     }
   }
 }
 
 .icon-close {
-  @extend %icon;
+  svg {
+    @extend %icon;
+  }
 }
 
 .modal-close {
